+++ src/elements/mod.rs
@@ -1,5 +1,7 @@
 mod b_box;
 mod b_button;
+mod b_notification;
 
 pub use b_box::*;
 pub use b_button::*;
+pub use b_notification::*;