+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 
 - Button Leptos component.
+- Notification Leptos component.
 
 ## [0.0.0-alpha.4] - 2024-01-29